Refactor component and file names, update color scheme and styles
- Renamed `TileGrid.tsx`, `TechTile.tsx`, `TechTileModal.tsx` to `00.TileGrid.tsx`, `01.TechTile.tsx`, `02.TechTileModal.tsx` respectively
- Updated import statements in `App.tsx` and related components accordingly
- Updated color scheme in `theme` object in `App.tsx`
- Updated background color in `AppBar` component in `App.tsx`
- Changed `color` prop value in `TileGrid` component in `App.tsx`
- Updated border color and hover color in `TechTile` component in `01.TechTile.tsx`
- Updated background color and responsive behavior in `TechTileModal` component in `02.TechTileModal.tsx`
- Removed unused components `CardGrid.tsx` and `TechCard.tsx`
